Add API `MediaSession.getBitmapDimensionLimit()`
Bitmaps used in media code could be scaled at various places. For example, in framework `MediaSession`, bitmaps larger than `MediaSession.mMaxBitmapSize` will be scaled down to fit within that dimension limit (see `MediaSession.setMetadata()`).

By having an API to query the dimension limit, it gives an app a chance to have the optimal dimension of the bitmaps in the first place. e.g. an app could request thumbnails/bitmaps that are within this limit, hence no scaling will be needed, and there will be no intermediate scaled bitmaps.

E.g. YouTube Music on pixel 7a requests many thumbnails of size 853x480, while `MediaSession.mMaxBitmapSize` is 840, this ends up with a scale down bitmap that is almost the same size, which could be unnecessary.

PiperOrigin-RevId: 805891255