Add `EventLoggerTest`
This adds some confidence that an upcoming refactoring won't
unexpectedly break the log formatting.

This doesn't test every possible permutation of more complex methods
like `onTracksChanged`, but it captures the general structure. The
future refactoring will move the 'detailed' formatting code intact, so
the part that we need to test is things like ensuring we open and close
the square brackets correctly.

PiperOrigin-RevId: 806254075