[demo-compose] Add a button to control the use of shutter
The actual timing will still be controlled by PresentationState and the presence of tracks, first frame being rendered and other relevant info. However, controlling `keepContentOnReset` can expose what a playlist of media items can look like when a shutter blocks the player Composable vs when the resize happens in front of the user.

PiperOrigin-RevId: 807710911