Reset frame adjustment at stream changes
This avoids getting a reset at a later mid-stream position
when any error accumulated far enough to trigger an reset.
In most cases it's preferable to time the reset with the
already existing stream change.

The desired behavior of onPositonReset to force the frame
adjustment reset at this point in time is already tested
by VideoFrameReleaseHelperTest

PiperOrigin-RevId: 807807533