Keep accumulated smoothed release time when encountering outliers
As long as the frame estimator is synced (all recent samples match
the estimated frame rate +-1ms), we keep accumulating a smooth
release time for as long it doesn't exceed a 20ms threshold to
the unadjusted release time. This basically ignores any audio
timestamp variations within this 20ms threshold and keeps adding
a the original frame timestamp diff to the previous relese time for
a smooth output.

As soon as we have a single frame time outlier >1ms off, we
completely reset this smoothing, which may cause a frame freeze or
drop if we already accumulated some difference to the unadjusted
timestamp. Even though we can no longer use the average frame rate
estimation, it's still better to keep the existing smoothed value
and keep adding to that for as long as it's within the same 20ms
threshold. This allows to kepp a smooth output even for occasional
small outliers in the media timestamp.

PiperOrigin-RevId: 807808846