Forward surface size changes between session and controller
This has been blocking demo-session from having some effects applied. CompositionPlayer required the surface size and the binder call only supported setting of the session. MediaController and ExoPlayer have local notification of surfaceSizeChanged to listeners, but this information would still get lost when communicating with the session.

Instead of creating a new method on the Player, we create a custom surface holder that combines the knowledge of the surface and resizing.

PiperOrigin-RevId: 808127742