Disable resolving asset list when manually changing ads
In case a user manually changes the state of an ad
that has an unresolved asset list, the asset list is
ignored in favor of the manual change.

In case the asset list resolution was loading the asset list
at the very moment of the manual change, the asset list load
is intercepted and reported as cancelled when it arrives.
This ensures to have a listener call complementing the
`onAssetListLoadStarted()` call that was reported when the
asset list started loading.

PiperOrigin-RevId: 808154035