Add emergency wakelock release if internal thread is unresponsive
The wake lock is released on the playback thread (to not call the
potentially blocking binder operation on the main thread). However,
we must not leak the wake lock, so when the player is stopped or
released after getting into a stuck state where the internal player
thread is no longer responsive, we should resort to releasing the
wake lock on the main thread (which is always available).

PiperOrigin-RevId: 811323105