Remove local vsyncOffsetNs variable
Having a separate variable suggests this is an independent
value from vsyncDurationNs, even though it's a fixed percentage
of it. It's clearer what we are tracking by only having one
variable that can change in the code. It adds the same
calculation for every frame, but the single integer
multiplication and division should be really cheap compared
to everything else that's happening.

PiperOrigin-RevId: 811346679