Default wake mode to WAKE_MODE_LOCAL
This enables wake lock handling in ExoPlayer by default. There is
no additional cost to holding a wake lock if the app or the audio
system already have another wake lock at the same time. Holding
one by default helps with issues during background playback in
particular for periods of time where the audio system is not holding
a wake lock for us (e.g. during buffering while no AudioTrack is
active).

The main risk are leaking wake locks where they are not needed.
ExoPlayer now has logic to detect any stuck state in which a wake
lock would be hold but no observable progress is made. It also
has a fallback mechanism to release the wake lock even if the
playback thread is blocked or gone. This means we can now more
safely enable it by default. If any of the main safeguards is
disabled, we keep the default at its previous WAKE_MODE_NONE.

PiperOrigin-RevId: 811409274