Guard codec buffer access by additional error checks
The buffers obtained from MediaCodec can be invalidated
after codec errors. Writing to an invalid buffer will cause
a native crash that cannot be avoided from Java. To fully
prevent access while the buffer is invalidated, this change
checks for any pending errors before starting to access
the buffer and uses the same lock as the error callback
to ensure the error callback can't run while the buffer
is being written to (as this happens on another thread).

PiperOrigin-RevId: 813167685