Move DisplayHelper logic into VsyncSampler
The sole purpose of DisplayHelper is to get the vsync duration,
so it logically belongs into the vsync sampling process. Combining
them into one helper class allows some code cleanup.

In addition, a future change may move to VsyncCallback from API 33,
where the vsycn duration is obtained through a single callback and
would need to be in VsyncSampler anyway.

PiperOrigin-RevId: 813175321