Use VsyncCallback from API 33
This can replace the previous FrameCallback. The advantage is
that we get the vsyncDuration as part of the same callback
and the value has a better resolution at nanosecond level than
the estimate we used so far from the screen refresh rate.

We still listen to display refresh changes to trigger a new
sample as soon as possible.

This change moves the vsync duration further away from what
Robolectric can fully simulate, so we need to adapt the test
to set the vsync duration based on the actual samples, matching
the real VSyncCallback behavior. One test also needed to adjust
the vsync variation in the test setup because it was way too
big to be realistic.

PiperOrigin-RevId: 813247055