add ShuffleOrder.cloneAndSet() to provide startIndex to shuffle order
Shuffle order may want to ensure that the song that's going to be played
(was chosen by user) is also the first song in shuffle playlist, as that
makes sure all other songs will be playing afterwards. If first song gets
random position in shuffled playlist, playback may stop before playing
every song as it will start in the middle of shuffled list.