Enable core desugaring for the IMA extension
This was missed as part of https://github.com/androidx/media/commit/653538e63e6bbb01ea539cab60817ea5dc90e148
because I only tested by building the main demo app, and thought adding
the config to the demo app was enough. This was shown not to be the
case when building the IMA extension instrumentation tests.

PiperOrigin-RevId: 817113247