Enable core desugaring for the proguard test module
This is needed because <unknown commit>
moved some Java 8 stream-based code from a test to "non test" code
(scare quotes because really this whole module is a test) - and
therefore the desugaring lint fails.

This module only contains tests, and isn't depended on by apps, so
this doesn't force the config on anyone else.

Also create a shared desugaring version constant now we're referring
to it in 3 places.

PiperOrigin-RevId: 817116528