Add thumbnail timestamp extraction to MatroskaExtractor
The heuristic finds the keyframe with the highest average bitrate within the first 20 chunk indices or the first 10 seconds of the video, whichever comes first. This assumes that a higher average bitrate over a chunk corresponds to more visually complex and thus make for better thumbnails. The timestamp is propagated via a `ThumbnailMetadata` entry in the track's `Format`.

Note: We delay sending `format` to the `TrackOutput` until we find the `Cues`.

Test: Added a media sample with multiple keyframes to validate the heuristic. The sample was created by forcing a GOP size of 10 on existing `sample.mkv`.
PiperOrigin-RevId: 817720766