Prevent Dolby Vision codec reuse when the profile changes
This change adds a workaround to force a codec re-initialization when switching between Dolby Vision formats with different profiles.

Note that there is fallback logic to the underlying encoding type when the display does not support dolby vision. This fix makes sure that in such a scenario the alternative mimetype codec is not reused.

PiperOrigin-RevId: 818696219