Fix scrubbing mode for DASH thumbnail playback
Scrubbing mode provides an improvement of skipping intermittent seeks so that scrubbing manifests a "frame progression" rather than rendering only when the user "stops seeking". This improvement currently relies on a `onVideoFrameAboutToBeRendered` callback conveying that a seek has been successful. The Image Renderer does not support this and so if scrubbing a thumbnail track then the player will not render any thumbnails until scrubbing mode ends.

This fix disables the "skip intermittent seeks" feature while scrubbing an image track so that images will still be rendered when the user "stops seeking". A future fix adding a "seek was successful callback" for images will re-enable the "skip intermittent seeks" feature.

Issue: androidx/media#2815
PiperOrigin-RevId: 820620189