Tweak some AudioOutput.FormatConfig fields
The OffloadMode is not needed (and doesn't really make sense here either),
and can be replaced by a boolean. For offload and tunneling, rename the
parameters to also indicate that this is an intention for the output
("enable" not "is). And lastly, annotate some of the parameters with a
note that they are not expected to influence the output of getFormatSupport
(as they are mostly relevant to getOutputConfig).

PiperOrigin-RevId: 820756997