Remove `equals()` and `hashcode()` impls from `PlayerId`
Both these implementations tie break based on `Object.equals()` on a
unique object that is constructed in the `PlayerId` constructor.

This means that `equals()` implements reference equality, which is
already provided by `Object.equals()`, so no need for these overrides.

This type is expected to always support reference equality, so this CL
also adds a test to enforce this.

PiperOrigin-RevId: 822061331