More clearly disallow C.INDEX_UNSET in PositionInfo.mediaItemIndex
It was never meant as an allowed value. More clearly catching it early
will help to uncover bugs at the source.

Also update some logic in MediaController where this value can
currently be used accidentally when setMediaItems is called with
an empty list.

The same is true for periodIndex, which we can assert more
clearly and fix issues with empty Timelines in ExoPlayerImpl
and SimpleBasePlayer where we accidentally set the wrong
value.

PiperOrigin-RevId: 823028361