Add method to round unset side in Presentation
When we downscale from 1080p to 480p, a 1 pixel wide black bar appears. This is because in Presentation we round 1080x1920 to 480x853, but then when encoding we letterbox this in another Presentation of size 480x854 (as this is the size supported by the encoder). Adding this method means we can round to the correct size in a single Presentation effect,

PiperOrigin-RevId: 828441212