Mp3Extractor: Align synchronizing and sniffing limits at 128kB
This value was split when sniffing was first introduced in <unknown commit>

But at that time, the sniffing order was fixed for **all** files, so it
was important that each sniffing implementation was quite conservative
in how much data it consumed, in order to not delay playback start-up
for files "later" in the sniffing order.

Since then, extractor sniffing order has been optimized to prefer
extractors that match the URI's file extension or MIME type (from a
response header):
* File extension: <unknown commit>
* MIME type: <unknown commit>

This optimization means that each individual extractor impl no longer
has to be so conservative in its sniffing implementation. The effect of
increasing the sniffing limit in `Mp3Extractor` today is:
1. For well-formed MP3 files with no/minimal garbage at the start,
   sniffing continues to return early when a sync byte is found - no
   change in behaviour.
2. For non MP3 files with a recognized file extension or MIME type,
   their preferred extractor will be sniffed before `Mp3Extractor`, so
   no change in behaviour.
3. For poorly-formed MP3 files with more than 32kB (but less than 128kB)
   of garbage at the start, playback previously failed and will now
   succeed.
4. For non-MP3 files with no recognized file extension or MIME type
   that are successfully sniffed by an extractor later than MP3 in the
   [default order](https://github.com/androidx/media/blob/05227c9591c6133abf4540e5460396919197ff45/libraries/extractor/src/main/java/androidx/media3/extractor/DefaultExtractorsFactory.java#L108-L133)
   - this will now take slightly longer, and involve
   reading 128kB data instead of 32kB (in order for the MP3 extractor
   to reject it). This is extractors for the following formats:
     * AVI
     * MIDI
     * Images

Issue: androidx/media#2713
PiperOrigin-RevId: 828884642