Disable stuck-while-suppressed detection in scrubbing tests
Scrubbing suppresses playback, and in a Robolectric test time can
auto-advance very quickly, which can easily exceed the default
timeout.

A couple of tests in `ExoPlayerScrubbingTest` were triggering
stuck-while-playing, so that's disabled for those tests.

Many tests in `ScrubbingPlaybackTest` use a blocking renderer
implementation which interferes with the stuck-without-ending
detection, so that is disabled for these tests.

#cherrypick

PiperOrigin-RevId: 828982225