[ui-compose-material3] Add text progress Composable
The Composable underneath is `BasicText` which is not smart enough to ensure that dynamic text, like video timestamp, doesn't cause layout to reflow.

Usually, there are a couple of ways to ensure the width remains constant:
* `Monospaced` font (every character has the same width)
* `fontFeatureSettings` parameter of the `TextStyle` (each digit has the same width, normal characters remain proportional)
* `TextMeasurer` (measure the widest "worst-case" string like 88:88, apply that to all other numbers)

The default font on Android (Roboto) is well-designed, and the width differences between its numerals are subtle to the naked eye. What is making it harder to test in the demo app is the flexible `Spacer` that takes up the *remaining* space.

Ideally, adding the tests that check the width of various strings **would** show that, as of now, the implementation is width-independent and robust. The problem is that the Android test environment is intentionally designed to prevent this kind of test from working. It does **not** use the standard "Roboto" font, but a basic, non-proportional test font (i.e. all characters have the same width). This is done to prevent visual inconsistencies and flakiness in screenshot and UI tests. In fact, loading custom fonts in this environment is a known difficulty, so no matter what FontFamily we specify, the test environment will likely override it with a font that has fixed-width numerals.

Given all this, we shall still apply one of the fixes to ensure the solution will work on real world devices.

#cherrypick

PiperOrigin-RevId: 829452658
(cherry picked from commit ea2303e78b7d73685655c4dce1e48c4effefefa5)