Tweak line-feed regex to avoid exponential backtracking
Before this change, the `(.|\f)*` pattern can lead to exponential
backtracking because it has to try many different combinations. Using
the extra `+` makes the matching 'possessive', so once a match is
found no other alternatives are considered. This is safe for these
regexes, because the pattern in question is always at the end of the
match, so there's no trailing characters that might need to be
'removed' from the match during a backtrack.

This is a follow-up to https://github.com/androidx/media/commit/2bd26008b42a4ce499a56e4d01074acfff343025 and https://github.com/androidx/media/commit/92677e93e2bbc26ec3db5e5b69bac4123229b9b9.

#cherrypick

PiperOrigin-RevId: 829457102