Preserve compressed bitmap data in controller legacy conversion
When MediaController connects to a platform session, it reads
bitmap data from the platform MediaMetadata and MediaDescription.
These structures may contain Bitmaps that need to be compressed
to be stored in Media3 structures like MediaMetadata. This
currently happens repeatedly, both for the same instance of
Bitmap, but also often for equivalent but not equal instances of
Bitmap (which are created by sessions sending updates over binder,
which create a new Bitmap instance every time even if the data
is the same).

This can be made more efficient by storing the already compressed
data in MediaDescriptionCompat and MediaMetadataCompat. And in
addition, we can attempt to preserve this data for new instances
if their original Bitmap is the same as the new one. Calling
Bitmap.sameAs is also not recommended for the main thread, but
is very fast (<1ms) for the Bitmap sizes returned from platform
media sessions, compared to the easily 100ms+ for every compress
operation.

PiperOrigin-RevId: 830410548
(cherry picked from commit 9f1ace0f55ea58a83b616c5e1ba9ca0c74036a24)