Remove ExoPlayerImpl.maskingPeriodIndex
This value was always hard coded to zero. This may be misleading
if the maskingWindowIndex is updated to a non-zero value because
it's not generally possible to have periodIndex=0 and windowIndex>0
in a player state. We can remove the value and since we are just
pretending to know the player state after preparation, use the
same placeholder value as the already existing maskingWindowIndex.

PiperOrigin-RevId: 830881591