[HLS] Properly handle getNextChunk for a playlist with no segments
When calculating the `segmentIndexInPlaylist` with the binary search (`stayInBound=true`), a corner case of `mediaPlaylist.segments` being empty also lead to the adjusted result of `segmentIndexInPlaylist = 0`. Then in the logic of picking up a `Part`, fetching a `Segment` may end up in `IndexOutOfBoundsException`. Thus we fix the `Part` picking logic into:

* When the playlist is VOD, there is no need to pick up a `Part`, then we bypass the `Part` picking logic;
* When the playlist is LIVE, and we are inside the live window:
  * When the `mediaPlaylist.segments` is not empty, use the old part picking logic: Depending on the `targetPositionInPlaylistUs`, either picking among the `Part`s of an existing `Segment`, or picking among the trailing parts. If we pick among the trailing parts, we should increase the media sequence by one as if we will be loading a "new segment";
  * When the `mediaPlaylist.segments` is empty, we can still try to pick from the trailing parts. If we pick a part, we don't need to increase the media sequence as we have hiddenly increased it by the `stayInBound` adjustment.

As a result, for VOD playlist, an empty `mediaPlaylist.segments` will lead to end of stream, and for LIVE playlist, it will lead to either loading a trailing part or refreshing the playlist.

Issue: androidx/media#2821
PiperOrigin-RevId: 830887686