Fix issue where previous PositionInfo uses unset window index
When receiving a PlaybackInfo update from the internal player
with a discontinuity, while ExoPlayerImpl has an empty timeline,
we currently use C.INDEX_UNSET as the placeholder window index,
which is not valid.

In addition, receiving these stale discontinuities while both
the current and new timelines are empty are misleading anyway, so
added a new check to avoid sending out these specifcally. The
fix to pass in the current masking window index is still needed
to be absolutely sure it's set in case there is a discontinuity
at the same time as a change from an empty Timeline to a
non-empty one (wasn't able to craft a test for this case).

PiperOrigin-RevId: 831406115