Correctly handle initial playlist if previously set items are empty
Checking if the list of media source is empty is wrong because these
media sources could all be empty, still resulting in an empty
Timeline. Check the Timeline directly to fully cover this case.

This bug was also the reason a later IllegalStateException in
evaluateMediaItemTransitionReason was triggered because the
state of ExoPlayerImpl and ExoPlayerImplInternal diverged,
causing this unexpected media item transition without any
other signal.

PiperOrigin-RevId: 831785624