Prioritize CBR seeking over index seeking for MP3s
This change alters the seeker selection logic in `Mp3Extractor` to provide a better default user experience for MP3 files that do not contain seeking metadata (e.g., a Xing or VBRI header). MP3 files without any VBR specific metadata are assumed to be Constant Bitrate (CBR), reflecting the evolution of the MP3 specification, where CBR was the original format and Variable Bitrate (VBR) support was added later and requires explicit marking.

Previously, if `FLAG_ENABLE_INDEX_SEEKING` was set, the extractor would create an `IndexSeeker` for such files. This provides accurate seeking, but only after the entire file has been scanned to build the index, leaving the media unseekable for a potentially long time.

With this change, the logic is updated to default to a `ConstantBitrateSeeker` when no other seeking metadata is available. This provides immediate but potentially less accurate seeking capabilities. The `IndexSeeker` is now used as a fallback if the primary seeker (from metadata or the CBR assumption) is not seekable, for instance, when dealing with a stream of unknown length.

Fixes: Issue: androidx/media#2848
PiperOrigin-RevId: 831861350
(cherry picked from commit f640c257030ed4021c9cde20df4c3006eab3ab66)