FSP: fix output capacity when flushing with texture listener
Texture listener is only used for packet consumer.
This fixes an issue where seeking with packet consumer
occasionally deadlocks

There is a race condition in ChainingGlShaderProgramListener (the inputListener).

onFlush() schedules videoFrameProcessingTaskExecutor.submit(producingGlShaderProgram::flush)
to run later, but the capacity in frameConsumptionManager is updated immediately.
The producing GLShaderProgram may produce a frame to frameConsumptionManager before
 the flush command is received.
 This would output a before-flush frame after the flush finishes.
 flushFinished runs inside an invoke() block, after the flush signal has propagated to
  the first shader program. It is safe to call onReadyToAcceptInputFrame in flushFinished.

PiperOrigin-RevId: 831875927