Use default start position for live pre roll interstitial
Given the HLS playlist sets the resume offset accordingly,
this change makes the pre roll ad group start and resume
at the default position of the live window.

It's the responsibility of the server to either set an explicit
resume offset or when omitting the resume offset, provide an
explicit duration of the interstitial in the playlist or the
asset list.

#cherrypick

PiperOrigin-RevId: 831911907
(cherry picked from commit a12d713ad362f103fa76f3bedf49fb34d1ea8ef7)