Fix potential NPE in DefaultPlaybackSessionManager
When handling timeline updates or position discontinuities, we
may remove the current session and later attempt to access it
again if we haven't explicitly cleared the currentSessionId.
This could happen if this currentSessionId has not been marked
as isCreated. Part of this change guards against that gap by
more clearly clearing the currentSessionId in any case.

This whole situation shouldn't happen though because every
situation where we set the currentSessionId should also ensure
that this session is marked as isCreated. One way to reproduce
an issue is to generate data for a not-created placeholder session
and then making this session the current one by matching it
with an event for a different window index. This would hit one
of the early returns in updateSession before the session is
marked as isCreated. To prevent this root cause, we can make
sure to not match such sessions in belongsToSession.
Additionally, the check in updateSessions for the sequence number
should only be used if there actually is a sequence number
to check (a bug that got accidentally introduced by https://github.com/androidx/media/commit/e0191ddded1754cc31b025493970240aeb08a46e)

Issue: androidx/media#2885
PiperOrigin-RevId: 832263214