Avoid retrying to load excluded playlists as the first primary playlist
In `DefaultHlsPlaylistTracker`, we have a `FirstPrimaryMediaPlaylistListener` which handles the load errors of the first media playlist and applies exclusion according
to the `LoadErrorHandlingPolicy` before the first media period has been created and prepared. It actually implements `PlaylistEventListener`, and according to the javadoc of `PlaylistEventListener.onPlaylistError`, the return boolean value suggests whether the playlist will be excluded from future loads, and then is used to instruct the `Loader` whether to retry to load the failed playlist (if exclusion failed, we retry the failed playlist, otherwise we do not retry).

However, the `FirstPrimaryMediaPlaylistListener` always return `false` for `onPlaylistError` even the failed playlist url has been excluded and we have selected a new primary url - https://github.com/androidx/media/blob/main/libraries/exoplayer_hls/src/main/java/androidx/media3/exoplayer/hls/playlist/DefaultHlsPlaylistTracker.java#L928. The linked `excludedPlaylist()` method returns a boolean value suggesting if the regarding url is primary despite of being excluded, which I further interpret it as "this url used to be primary, and we cannot find another primary url, and it remains primary", thus this return value being `true` suggests that the exclusion failed, otherwise it suggests that the exclusion succeeded, then we shouldn't retry the failed url as it is no longer primary.

PiperOrigin-RevId: 833297663