Remove extra `onReadyToAcceptInputFrame()` call
There's one extra call to `InputListener.onReadyToAcceptInput()` invocation in
`queueInputFrame()` method, on the first frame.

The issue is, `InputListener.onReadyToAcceptInput()` is
called twice for the first frame:

- Once in the `queueInputFrame()` method,
- Once when the consuming shader program calls `onInputFrameProcessed()`
  - This calls to `BaseGlShaderProgram.releaseOutputFrame()`, which
    calls `InputListener.onReadyToAcceptInput()`

This disturbes the capacity counting in `FrameConsumptionManager`.

This issue is not easily reproducible when the `DefaultFrameDroppingEffect`
connects to `DefaultGlShaderProgram` implementations, but it'll be obvious
when connected to raw `GlShaderProgram` implementations, like
`DebugViewShaderProgram`.

A test is added to ensure `DefaultFrameDroppingEffect` works.

PiperOrigin-RevId: 833369945