Fix asset list message scheduling
For live, the public content position with which to query
the pending asset lists must be converted to a period
position with the windowPositionInFirstPeriodUs of the
window with which the content position was calculated.

Thus for a live stream, this conversion of the content
position must always be done with the timeline that has
been used by the player when returning `getContentPosition()`.

#cherrypick

PiperOrigin-RevId: 833434897