Extract MediaExtractorCompat internal implementation to separate class
This change moves the core logic for media extraction from `MediaExtractorCompat` into a new `MediaExtractorCompatInternal` class.

Following this change, `MediaExtractorCompat` becomes a thin wrapper that delegates all of its calls to an instance of `MediaExtractorCompatInternal`.

This is a preparatory refactoring to enable the future introduction of `MediaExtractorCompat` in the new `:media3-inspector` module. It allows both the existing (and soon-to-be-deprecated) class and the new one to share the same underlying implementation during the migration period without creating circular dependencies. This pattern is similar to the one used for `MetadataRetriever` in https://github.com/androidx/media/commit/a1bc8de4da490c7b0f4d857de7433b51260bcd1c.

#cherrypick

PiperOrigin-RevId: 833785780