Handle edge case of command change not yet propagated to button state
The button enable state depends on the available commands. When they
change though, the isEnabled state is only updated for the next
recomposition. This means that calling onClick() may happen at a point
where the button is still "enabled", but the command can no longer be
called.

The method contract is to be callable when isEnabled==true, so we
should change all assertion to check this explicitly. Because of
this edge case, we also need to verify the relevant player commands
are available just before calling the respective player method.

#cherrypick

PiperOrigin-RevId: 833841451
(cherry picked from commit 348a3846d79ab3d63422dfb7ecdd03fb8fd695ac)