Run different API entry points in TransformerAndroidTestRunner
This change refactors the `TransformerAndroidTestRunner.run` methods to call the respective `transformer.start` method depending on the type of `mediaInput` passed (`MediaItem`, `EditedMediaItem`, or `Composition`). This makes it more robust for testing the different API entry points of `Transformer.start`.

This is a part of the effort to migrate to specifying track types for `EditedMediaItemSequence`.

PiperOrigin-RevId: 833843043