Align ProgressStateJob lifecycle to observe operation
Currently, the lifecycle of the update job is only loosely
coupled to the observeProgress method call. It's started
by this call, but not stopped if the observation goes out
of scope. It may also be started (without ever calling
observeProgress) by updating the tick interval.

To avoid surprising effects or overruning update jobs,
we can more clearly align the job lifecycle to run
for at most as long as the observeProgress method.

#cherrypick

PiperOrigin-RevId: 833857639